Added menu or in game detection
This is a complete hack at the moment. We check if a certain part of the
emulator's heap is allocated with a menu struct (?). Part of that struct
has an address list to jump to other menus (I think). The code checks if
the beginning of the address list starts with 0x81 (probably a valid
addr). Again, this is a hack so it will need to be fixed later.